Add id getter setter in AbstractEntity to avoid duplicate data
