+++ index.php
@@ -2,6 +2,7 @@
 declare(strict_types=1);
 
 use App\Core\Kernel;
+use App\Entity\UserEntity;
 
 date_default_timezone_set('Europe/Paris');
