+++ src/entity/AbstractEntity.php
@@ -1,3 +1,19 @@
 <?php
 
 namespace App\Entity;
+
+class AbstractEntity
+{
+    protected int $id;
+
+    public function getId(): int
+    {
+        return $this->id;
+    }
+
+    public function setId(int $id): self
+    {
+        $this->id = $id;
+        return $this;
+    }
+}
