+++ src/entity/PostEntity.php
@@ -4,7 +4,7 @@ namespace App\Entity;
 
 use DateTime;
 
-class PostEntity
+class PostEntity extends AbstractEntity
 {
     private int $id;
     private array $comments;
